Fix caught exception leaking through `except` scope (#715)
Summary:
As reported in https://github.com/facebook/pyrefly/issues/154, if a caught exception is bound to a name `e` in a `except` block, the caught name `e` should be out of scope after the corresponding `except` scope.

This is reflected in the spec as follows
https://docs.python.org/3/reference/compound_stmts.html#except-clause:

> When an exception has been assigned using as target, it is cleared at the end of the except clause.

This PR aims to model this effect by marking the caught exception name as uninitialised at the end of the current `except` block.

Closes https://github.com/facebook/pyrefly/issues/154.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/715

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

Rollback Plan:

Reviewed By: stroxler

Differential Revision: D78898967

Pulled By: yangdanny97

fbshipit-source-id: 2d737d8aebbfafb6efae3e6241e32f488e13ae7a