Do not assume variable is uninit with only annotations (#751)
Summary:
Closes https://github.com/facebook/pyrefly/issues/246.

When we encounter an annotated statement without assignment, e.g. `x: int`, we assume that the variable is uninitialised if it's not occuring in a class. This is not universally true, as pointed out by the example reported in the issue.

Instead, this commit proposes to check for the current flow style of the variable being annotated using `get_flow_style` function. The function is able to check for initial use and give the correct uninit flow style, but otherwise returns the existing flow style (e.g. if the variable has already been written to).

Also fixes a small typo nearby.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/751

Reviewed By: rchen152

Differential Revision: D79099201

Pulled By: stroxler

fbshipit-source-id: 71af219baea6a3f3fc89737691cb030471186107