TypeDisplayContext understands if types are toplevel
Summary:
I would like `TypeDisplayContext` to understand nesting correctly: certain types could be displayed longer if they exist alone but should be displayed more concisely if they are part of another type.

See the next diff for an implementation that would use this.

How can we implement this?
**(this approach): store is_toplevel on the stack**
This should work the best in theory. But unfortunately a lot of nested calls call `display` again, resetting the `is_toplevel` indicator. In this diff, I've made a `display_internal` that should be used for nested calls, but I worry we will accidentally use `display` instead internally.

I've looked into ways to prevent this:
- visibility: rust only supports pub as a means to make a method visible externally and internally at the same time. we can't solve this with basic visibility operations
- wrap the class: can we wrap `TypeDisplayContext` so we can expose a different `display` function vs internal callers? (I think this is likely the best approach)

**(other approach): add `is_toplevel` to `TypeDisplayContext` struct**
By adding it to the struct, we can mutate it to false after any function call. But this approach requires making `self` mutable which does not work well with the current setup.

I'm looking for comments on these approaches / if there are other ideas

Reviewed By: SamChou19815

Differential Revision: D79096954

fbshipit-source-id: 8b119c5299a6f8d6a87b647cb0ec607c5703cd9b