Use `YieldLink` for linking yield calculations
Summary:
The `UsageLink` key is TextRange based, so we cannot have
overlapping ranges.

It's pretty easy to guarantee this in `stmt.rs` because statements
are non-overlapping, so you can just read down the cases and see that
ranges never overlap.

But it's much harder to guarantee that `stmt.rs` and `expr.rs` never
produce overlapping ranges - in particular a `del` statement can have
multiple targets so we *have* to key on something other than the full
statement range, which makes it tricky to use a single key type to fix
a crash on `del yield, "x"` (which panicked because both the first del
binding and the yield binding tried to use the `yield` keyword range
for their `UsageLink`).

The easy approach is to just have `yield` links, which are the only
links in `expr.rs`, use a separate key variant so that the overlap is
no problem. It's also now easy to just skim the code and see that overlapping
ranges are not possible, since `stmt.rs` and `expr.rs` no longer try to use
the same key variant.

Fixes https://github.com/facebook/pyrefly/issues/777

Also fixes an `assert` case that would have panicked for the same reasons

Reviewed By: ndmitchell

Differential Revision: D79098815

fbshipit-source-id: 7583360e7a8b2d92ed830243fec474b1d345116e