Calculate `dataclass_metadata` outside of base class metadata loop
Summary:
refactortractor

When calculating class metadata, we have a big loop over the base class list of the given class. Although by name it seems like the loop is only responsible for turning a list of `BaseClass` into a list of corresponding `(ClassType, ClassMetadata)`, the implementation does so much more than that -- we have a whole bunch of states declared outside of the loop but getting mutated within the loop, which might be fine when we first started but as time went on and the loop became big & complicated, I think the readability/maintenance cost of mentally tracking all those individual states now far outweighs the tiny perf benefit of calculating all states with a single loop iteration.

This diff proposes that we lift some of the calculation outside of the "big loop", so the latter can be simplified, which in turn prepares for my real goal of eliminating the dependencies between the "big loop" and base class targs.

We could easily lift out the logic of finding `dataclass_metadata`, since the info can be 100% determined by iterating over the *result* of "big loop" (as opposed to calculating the info during the "big loop"). For this change to be 100% semantic preserving, I am assuming that dataclasses are never tuples or typed dicts at the same time, which feels like it must be the case.

Reviewed By: stroxler

Differential Revision: D79119719

fbshipit-source-id: ba52a34a0ad76b84d8c6462f9d6fcd57c31ad4ed