Calculate `has_generic_base_class` and `has_typed_dict_base_class` outside of base class metadata loop
Summary:
code_refactoring

When calculating class metadata, we have a big loop over the base class list of the given class. Although by name it seems like the loop is only responsible for turning a list of `BaseClass` into a list of corresponding `(ClassType, ClassMetadata)`, the implementation does so much more than that -- we have a whole bunch of states declared outside of the loop but getting mutated within the loop, which might be fine when we first started but as time went on and the loop became big & complicated, I think the readability/maintenance cost of mentally tracking all those individual states now far outweighs the tiny perf benefit of calculating all states with a single loop iteration.

This diff proposes that we lift some of the calculation outside of the "big loop", so the latter can be simplified, which in turn prepares for my real goal of eliminating the dependencies between the "big loop" and base class targs.

The more obvious piece of data we could calculate outside would be the flags that track whether a class has generic base or whether it is typed dict. These can be calculated *before* entering the big loop by simply examining the kind of `BaseClass`.

Reviewed By: stroxler

Differential Revision: D79119720

fbshipit-source-id: 351c7e55defc6596398fba180a9461b600bd6cab