Caculate `is_typed_dict` outside of base class metadata loop
Summary:
refactor

When calculating class metadata, we have a big loop over the base class list of the given class. Although by name it seems like the loop is only responsible for turning a list of `BaseClass` into a list of corresponding `(ClassType, ClassMetadata)`, the implementation does so much more than that -- we have a whole bunch of states declared outside of the loop but getting mutated within the loop, which might be fine when we first started but as time went on and the loop became big & complicated, I think the readability/maintenance cost of mentally tracking all those individual states now far outweighs the tiny perf benefit of calculating all states with a single loop iteration.

This diff proposes that we lift some of the calculation outside of the "big loop", so the latter can be simplified, which in turn prepares for my real goal of eliminating the dependencies between the "big loop" and base class targs.

Now that D79119720 factors out the logic of determining whether a class contains a base class typed dict or not, this diff takes it one step further -- to figure out whether a class is typed dict or not, we examine whether it contains a base class typed dict first, and if not we examine whether any of its base class metadata says if the base class is a typed dict.

Reviewed By: stroxler

Differential Revision: D79119721

fbshipit-source-id: 3c8d88b7ebfcc1388e0bd15d8efb3f49f8783884