Don't treat assigning to an attribute in a method body as re-declaring it
Summary:
If the first time we see an attribute is when it is assigned to in a method body, we want to treat that as the attribute declaration. But if the attribute is inherited, then the assignment is just an assignment, not a re-declaration. Note that this applies only to an unannotated assignment (`self.a = b`); an annotated assigment (`self.a: X = b`) is still treated like a declaration.

Fixes https://github.com/facebook/pyrefly/issues/753.

Reviewed By: stroxler

Differential Revision: D79117575

fbshipit-source-id: 6afa7dbf3da00a17c9437ba93ed9108973bf0cd6