Calculate `named_tuple_metadata` outside of base class metadata loop
Summary:
quick_refactoring

When calculating class metadata, we have a big loop over the base class list of the given class. Although by name it seems like the loop is only responsible for turning a list of `BaseClass` into a list of corresponding `(ClassType, ClassMetadata)`, the implementation does so much more than that -- we have a whole bunch of states declared outside of the loop but getting mutated within the loop, which might be fine when we first started but as time went on and the loop became big & complicated, I think the readability/maintenance cost of mentally tracking all those individual states now far outweighs the tiny perf benefit of calculating all states with a single loop iteration.

This diff proposes that we lift some of the calculation outside of the "big loop", so the latter can be simplified, which in turn prepares for my real goal of eliminating the dependencies between the "big loop" and base class targs.

`named_tuple_metadata` can now be 100% calculated *after* the big loop (as opposed to getting calculated within the big loop).

Reviewed By: stroxler

Differential Revision: D79134767

fbshipit-source-id: 00ac4bbee770ddb98d5b26c4a0fae8316d2d1f5f