Simplify updating based on the config
Summary:
Now have the caller call invalidate_config. Before we invalidated the config even if we didn't modify anything, and we invalidated it once per scope in some cases. Now we invalidate at the end.

We now do also invalidate config if you turn off type checking, which isn't necessary as we only need to update the memory files, but this should be minor additional work for a simpler API.

Reviewed By: SamChou19815

Differential Revision: D79121561

fbshipit-source-id: 99f8b0349f3520588891ddc3cb1272f03bed40de