Do not use Forward downstream of a merge
Summary:
If we are able to merge all branches to a single key (i.e. the name wasn't
impacted by branching), then ideally we should just use the original idx.

Using a forward isn't very good because it's not composable - yes, if you
look narrowly at the local code it will resolve the same. But two equivalent
Forward idx's will be completely different, which means that for more complex
code with layers of branching we'll produce useless phi nodes on
outer layers when inner layers have merges that turn into (non-equal) Forwards.

That's silly - it messes up some semantics (nested control flow breaks first-use
tracking) and also leads to expensive solve-time compute because the Phi nodes
will call `self.unions` which has to allocate and then deduplicate a bunch of
type_info's that we should have known were identical ahead of time.

The code is slightly uglier now since we have to insert the phi_idx binding and
determine the downstream binding in a less-coupled way, but the benefits of this
should be pretty substantial.

Fixes:
- the bug (no issue filed) where nested no-op control flow breaks first-use
  based type inference
- https://github.com/facebook/pyrefly/issues/657

It also eliminates a var leak through an error message on another bug, although
it doesn't fix the bug itself.

Reviewed By: ndmitchell

Differential Revision: D79096377

fbshipit-source-id: e764ab477ed93058de7310145530f6f772684739