Eliminate `try_lookup_instance_method`, reduce cloning
Summary:
It had exactly one use, which was in fact just a less
efficient version of `instance_as_dunder_call` that forced
an extra clone!

We introduced this because `instance_as_dunder_call` was in
the wrong module, so the people introducing the duplicated
functionality had no realistic way to even knew the logic already
existed.

Reviewed By: rchen152

Differential Revision: D79184658

fbshipit-source-id: 33d2430f15de7b8243ea0fd32b907c91be3c33a8