Split base class metadata loop further into 2 loops
Summary:
banana_split_shipit

There were some checks within the "big loop" and the only reason they were there is because they want access to the `range` of the base class. By preserving the range into the result of the "big loop", we could now lift those checks out of the "big loop" and further simplify it.

Reviewed By: ndmitchell

Differential Revision: D79136131

fbshipit-source-id: 302144e408bf98a42f70a36ec47be6ea9140e61e