Calculate `has_base_any` outside of base class metadata loop
Summary:
thisguyrefactors

When calculating class metadata, we have a big loop over the base class list of the given class. Although by name it seems like the loop is only responsible for turning a list of `BaseClass` into a list of corresponding `(ClassType, ClassMetadata)`, the implementation does so much more than that -- we have a whole bunch of states declared outside of the loop but getting mutated within the loop, which might be fine when we first started but as time went on and the loop became big & complicated, I think the readability/maintenance cost of mentally tracking all those individual states now far outweighs the tiny perf benefit of calculating all states with a single loop iteration.

This diff proposes that we lift some of the calculation outside of the "big loop", so the latter can be simplified, which in turn prepares for my real goal of eliminating the dependencies between the "big loop" and base class targs.

`has_base_any` is set to True in 2 cases: either the class itself has an `Any` base, or the base class of the class has an `Any` base. We can always compute the latter after the big loop (in the same fashion as D79119719 or D79134767). But the compute the former we'll need to track whether any of the base class is invalid or not. Previously we rely on `filter_map` which drops those invalid base classes immediately so subsequent logic can't really get the info. But we could easily change the logic to track whether the dropping happens (I'm using `partition_result()` here whether the err case is just a single unit -- happy to change if there's a more idiomatic way of doing it) and use it accordingly.

Reviewed By: ndmitchell

Differential Revision: D79134769

fbshipit-source-id: 5de8fbe7b7fe98855704b1b38dfbbd9ad224ce6a