Add minimal stub and unittest for ConfigDict
Summary:
Based on the writeup https://docs.google.com/document/d/1AVe4QknDMZingQa-KHlFNEbYVIp139UrUsB3tg3myoo/edit?tab=t.0  there are certain unsupported configs. This adds a unittest and stubs for one of them.

In this test, we're supposed to be able to tell that `x` is immutable.

mypy catches the error here and we don't.

Questions for me to think about after ramping up on dataclass implementation:

- is this reasonable typing behavior overall or is it pydantic specific? (pyright doesn't support this pattern either)

- If it is pydantic specific, where should the code live so it doesn't interfere with the existing implementation

Reviewed By: rchen152

Differential Revision: D79215663

fbshipit-source-id: 8b0210042408bb7dbe3d1debbb686eb9323d4325