Fix leaking Self type when treating a constructor as a callable
Summary:
When converting a constructor to a callable, we should be careful to replace Self with the class type. Leaking the Self type leads to issues because we (rightly) consider two Self types to be equivalent.

I was able to find a simple case of unsoundness (see linked issue).

Leaking a self type can also lead to confusion during self-substitution. If a Self type leaks into a type which is then Self-substituted with a different type, we will substitute both Selfs. This leads to the other test included in this diff.

Fixes https://github.com/facebook/pyrefly/issues/629

Reviewed By: rchen152

Differential Revision: D79213792

fbshipit-source-id: 4929c0fbb80a50b5e7268ce07cd0869118f841df