Make LSP send the exit notification
Summary:
We wait for the LSP to shut down, but that only happened because we were dropping the right connections in the right places which meant we got receive errors. The right way to shut down is by sending exit. So let's do that instead.

In a few diffs time it won't work unless we send exit, as we change the lifetime in ways that are irrelevant to real LSP, but matter for our test.

Reviewed By: SamChou19815

Differential Revision: D79197340

fbshipit-source-id: 275089535611dbce087274b132360ce35274e41c