Substitute in get_inherited_type_and_annotation
Summary:
It's always a bug to try to use raw parent class fields without substituting.
We're *very* careful of this in normal attribute lookup code, but as
the class field binding has grown into a massive codebase with a lot
of folks hacking on it it's easy to introduce new ways of doing
similar lookups and forget the subsitution.

In this case, it led to both extraneous errors in the class field trying to
validate it against an unbound quantified, and to the wrong type (an unbound
quantified) leaking out of the final result.

Easy fix: both the type and the annotation need to be substituted.

Reviewed By: rchen152

Differential Revision: D79201627

fbshipit-source-id: fd03487be849bd235b1a9b51dccf040c4263281b