Simplify loop_depth
Summary:
I was always a bug to loop over the scopes. This fixes weird behaviors when a function
is defined inside a loop - the outer loop shouldn't be modifying how we handle flow
inside the function.

I'm not sure why we were casting to u32, it's not saving us any RAM since we already had
a usize, so while I'm at it I just use usize.

Reviewed By: grievejia

Differential Revision: D79268043

fbshipit-source-id: c7b3be478eab2e88ecf25b46c7e04c3aeccb3dd7