Rewrite the fuzzy searching code
Summary: It was doing a clone_for_mutation, then throwing away the resulting mutations. That might mean you are in a state where you mutate the thing you cloned (e.g. remove deps), but mutate the underlying thing for the dependency (e.g. remove deps). That leads to the crash on reported in https://github.com/facebook/pyrefly/issues/741.

Reviewed By: SamChou19815

Differential Revision: D79343957

fbshipit-source-id: 1300b9efcfb1681f916a4542ac2d8c2d077e11f9