test_completion_basic: consistent builtins across platforms
Summary:
test_completion_basic now includes all builtin completions, but they differ across platforms. in the test, we can:

**create real builtin completions by generating a list of exports based on python version**
We would need to demand an entire new exports step for the `typeshed()` module - this seems like the best approach but I'm not exactly sure how to do that right now.

**enumerate the builtin completions per python version for the test**
this will get annoying as python versions update and this test updates.

**delete the test since other tests already cover it's behavior**
i'm considering this

**get the test to only check for the completion we care about**
this diff: we ignore the entire request after the first completion. we got lucky that Bar is the first completion, but we always add local completions first so i don't see this changing any time soon.

I chose option 4. I also removed an unnecessary part of this test since it tested the same thing twice.

Reviewed By: ndmitchell

Differential Revision: D79345667

fbshipit-source-id: ee8d05c55f53135a13b5c460fd6f34d81071c1b5