Don't even call set_memory if there are subsequent mutations
Summary:
If there are subsequent mutations we:

1. Changed the memory with set_memory.
2. Then didn't do a run.

Turns out that breaks an invariant that 2 breaks an invariant I want to assert on in the next diff, and 1 is redundant because everyone calls set_memory all the time, and we will get the update on the next go. So just skip the thing entirely.

Reviewed By: kinto0

Differential Revision: D79344375

fbshipit-source-id: eb4492142402ff7aa2057e20f068c87952feadf4