Rewrite invalidate_rdeps
Summary:
This code used the list of readable modules and deps before the code had run, but at this time we may have run a couple of steps, which may have updated that information. Therefore the data it was working on was stale. That can't be good, and indeed, I believe it was implicated in breaking some invariants that a susequent crash tripped over. But I haven't been able to replicate it in a test.

Happily, since this code was written, we now store real rdeps. So if you want to do a transitive closure of rdeps, rather than iterating over all modules, you can just do a transitive closure following the rdeps. Much simpler, and uses the current data, rather than the stale data. I hope this fix some bugs!

Reviewed By: kinto0

Differential Revision: D79358373

fbshipit-source-id: 293771eaa7467716c607df1bbec591a6c718661a