Avoid storing types in `BaseClass`
Summary:
abstractthingproviderfactorybasebuilderwrapper

The goal here is to avoid trying to turn expressions into types for base classes in class metadata computation, to avoid creating cycles between base class tparams and current class metadata.

The class metadata computation currently relies on computing `BaseClass` from base class expr first. And the `BaseClass` structure currently holds types instead of expressions, meaning that it basically forces an `expr_infer()` somewhere within its construction. This is bad because it in turn forces class metadata to transitively invoke `expr_infer()`.

This diff attempts to change the `Type`s in `BaseClass` to `Expr`s, so we could avoid the call to `expr_infer()` when calculating `BaseClass`. There is one check on multiple occurrences of `TypeVarTuple` that depends on those fields being `Type`s, but it can be moved out of `BaseClass` construction and into downstream logic. This move is beneficial in general anyway -- given that we construct `BaseClass` when solving both class metadata and class tparams, previously those errors will be duplicated but with the check moved out we no longer do that.

Now that `BaseClass` can be constructed syntactically, in subsequent diffs we should be able to calc in during the binding phase, which could in turn leverage the `SpecialExports` mechanism.

Reviewed By: stroxler

Differential Revision: D79377455

fbshipit-source-id: 7458b85d7bb6727c93fe57e19bbd57f758623b55