Fix instantiation of generic classes with overloaded constructors
Summary:
Before this diff, when we constructed a class, we first instantiated the class with fresh vars for each tparam, then called the constructor. If the constructor was overloaded, we would pin the var at its first use, even if the overload that caused the pin ended up failing.

This diff rewires the code to construct a class and defers the creation of fresh vars until right before handling a callable. In other words, we create seprate vars for each arm of an overload. We can then "choose" the vars which correspond to the matching overload only.

Another complexity is that constructors are not necessarily as single overloaded call. A class can define both `__new__` and `__init__`, and either one can provide types for the class tparams. In fact, the spec even requires that `__new__` be able to _partially_ instantiate a class, then completed by `__init__`.

To accommodate this, I needed to add some more complicated machinery to "generalize" the targs. This is the opposite of "fresh," where we take a Var back to a Quantified.

Finally, once we are done, any unsolved vars need to be dealt with. Other type checkers solve these vars to `Any`, but Pyrefly tries a bit harder to infer a type, so we leave a `Quantified` var which will get pinned at it's first use downstream.

Fixes https://github.com/facebook/pyrefly/issues/315
Fixes https://github.com/facebook/pyrefly/issues/384
Fixes https://github.com/facebook/pyrefly/issues/487
Fixes https://github.com/facebook/pyrefly/issues/764

Ref https://github.com/facebook/pyrefly/issues/680 - I wasn't able to reproduce this error before this change.

Ref https://github.com/facebook/pyrefly/issues/295 - this change fixes the reduced example from Zsolt, but the original question involves narrowing of Any on assignment.

Reviewed By: stroxler

Differential Revision: D79234124

fbshipit-source-id: 706747405fbeb15668b435d59968bc13b2832565