validate in memory before document diagnostics
Summary:
since we now [skip validating in memory if there's a subsequent mutation](https://fburl.com/code/bfqya0tr), we are actually skipping running the state before returning a textDocument/diagnostics response. side effect: we now are guaranteed a publishDiagnostics before returning a textDocument/diagnostics (what we would expect).

this is the opposite of that we saw in D79344416: in the current failure, we actually don't find any errors at first when we should.

doing this transaction might duplicate work, but this is low risk because textDocument/diagnostics is never used in production.

Reviewed By: ndmitchell, SamChou19815

Differential Revision: D79413266

fbshipit-source-id: 030dd7119359a06d501b18f7f934b683bb3437d6