Don't overwrite `FlowStyle` when binding names in legacy tparams builder
Summary:
legacy tparam builder overwrites flow style when bind names. I believe this is undesirable because it messes with `SpecialExports`, which relies on flow style to figure out where a name is imported from.

This is a quick-and-dirty fix following what was done in D78496802, where we perform a lookup before inserting the style. The flow style info in general seems to be in a messy state (iirc we run into this overwrite problem somewhere else and the solution was actually different), but I'm not quite ready to clean it up as I need to focus on D79410203.

Reviewed By: stroxler

Differential Revision: D79410202

fbshipit-source-id: eaf4f33e1249ef6f91c166e1223fc3cb10f407cd