Calculate `BaseClass` in binding phase
Summary:
This diff proposes that we convert exprs in the base class list to `BaseClass` right at the binding phase.

Previously, we were converting exprs to `BaseClass` at solving phase and relied on `expr_infer()` during the conversion. As mentioned in D79377455, this need to change if we want to break the class metadata cycle. Since D79377455, the conversion logic no longer need to dependency on `expr_infer()`, and this diff eliminates that dependency by rewriting the logic using the syntax-oriented `SpecalExports`, rather than any typing info.

Reviewed By: stroxler

Differential Revision: D79410203

fbshipit-source-id: dbb3ed506eca86f9d22b0a1f828e9e23518224a5