Fix favicon in subdirectories
Summary:
We used a relative URL for favicons, which means that we would fail to find a favicon if you navigated to pyrefly.org/sandbox or pyrefly.org/en/docs (confusingly, if you start at pyrefly.org and then click to navigate, the favicon is "sticky").

Use an absolute URL instead so it works on every page.

Noticed this because I had a ton of Pyrefly sandbox tabs open, and none of them had favicons :)

Reviewed By: rchen152

Differential Revision: D79471735

fbshipit-source-id: ceba60ecfed657862220dadf7307d0fe3d28b972