Replace a semantic token test with a gotodef test in D79385434
Summary:
I noticed that I used a semantic token test D79385434 to demonstrate the issue was fixed. Looking back at it, although the test works it does feel rather unintuitive and it also confused kinto0 a bit.

I believe the same issue is originally reported as a gotodef issue and therefore it's much more straightforward to use a gotodef test instead. Both tests point to the same issue, but gotodef may be easier to skim / understand.

Reviewed By: kinto0

Differential Revision: D79459284

fbshipit-source-id: eeda74cea7dbdd5db54e1ae4d0672786f5ad7a70