Make sure we promote literal string as well
Summary:
Fixes https://github.com/facebook/pyrefly/issues/822

Shows up a lot less frequently, but should have been promoted too. We had a test that we deliberately didn't promote it, but that was dubious as:

```
def f(x: Literal['test']):
    xs = [x]
    assert_type(xs, list[str])
```

So it's weird we do promote there, so this is consistent. You could argue for explicit human annotations we should propagate inside list, but today we don't anyway, so should change everywhere.

Reviewed By: kinto0

Differential Revision: D79467127

fbshipit-source-id: eb00ff2df9ea63b30445f1da0aad190ba16a5abf