Export class information for synthesized classes
Summary:
# Context
The `--report-pysa` option exports type information to files so it can be used to power Pysa's taint analysis.

# Problem
We try to export all classes, but we actually miss many classes. This is because we only traverse classes visible in the AST, but the type checker can actually create artifical classes. For instance, `Point = namedtuple('Point', ['x', 'y'])` will create a class that is not in the AST (no `class xxx` statement) but is present at runtime.

# Solution
Instead of using the AST to find all classes, use `bindings.keys::<KeyClass>` to get all visible classes from pyrefly.

Reviewed By: grievejia

Differential Revision: D78989035

fbshipit-source-id: 640e772fa1b60f51d7bddd21eb6db90cf053ad7a