Rewrite function definitions to handle type parameters that appear only in a Callable return type
Summary:
When a function returns a Callable that contains type parameters that don't appear in the parameter list, both pyright and mypy treat the return type as being a generic callable. This diff copies this behavior by putting a Forall around the function's return type when constructing the function type.

That is, if we have:
```
def f[T1, T2](x: T1, y: T1) -> Callable[[T2], T2]: ...
```
we would previously produce the type:
```
[T1, T2](x: T1, y: T1) -> ((T2) -> T2)
```
but now we produce:
```
[T1](x: T1, y: T1) -> ([T2](T2) -> T2)
```

Fixes https://github.com/facebook/pyrefly/issues/750.

Reviewed By: samwgoldman

Differential Revision: D79309544

fbshipit-source-id: 50715ec8dbe0992f6c200e0a8be972eeb3e01a39