Remove use_untyped_imports
Summary:
Pyrefly has support for PEP 695's [`-stubs` packages](https://typing.python.org/en/latest/spec/distributing.html#partial-stub-packages), along with checks for whether we can use those packages for typing information.

By default, we have this functionality disabled, since we can automatically infer the types of imported modules (or mark them as `typing.Any` when they're compiled), and because it could make for a confusing/irritating user experience. Because of this, I'm inclined to believe this is a rarely used feature, if I'm not the only one who has used it :).

This feature is also somewhat difficult to maintain as things are, and makes it really difficult to prefer .pyi files over .py files if they occur later in the search path or site package path. We can rearchitect how this works, but I honestly don't think it's worth keeping around. We can always come back to it if we decide it's worth having.

Due to the above reasons, I'm going to disable it, and eventually remove it completely from configs.

Reviewed By: grievejia

Differential Revision: D79480532

fbshipit-source-id: a948cb0c0e78777c497af67896b03c286628470e