Handle unions including TypedDict when inferring a type for a dict literal
Summary:
In the code for inferring a type for a dict literal, TypedDict hints are handled in a special case before the more general case of builtins.dict hints. This diff extends the TypedDict special case to handle TypedDict options in a union.

Note that the non-TypedDict case (https://github.com/facebook/pyrefly/issues/793) is still buggy; this diff doesn't make the situation there better or worse.

This is essentially an updated version of D74258897. I saw that that diff was shelved because it didn't solve the non-TypedDict case , but I want to get in this TypedDict-specific fix because it's affecting our ability to handle pydantic.ConfigDict.

Fixes https://github.com/facebook/pyrefly/issues/164.

Reviewed By: samwgoldman

Differential Revision: D79515920

fbshipit-source-id: 39bfa8012422b4cb7f97377ad3e44c5369105345