Add information about the parent of functions and classes
Summary:
In pysa, each class and function needs to be represented by a unique fully qualified name. See D78832811 for more context.

With the current design, pysa is responsible for building the list of classes and functions. This is problematic because pyrefly might synthesize functions and classes that are not visible in the AST. For instance, `P = namedtuple('P', ['x', 'y'])` creates a class `P`. This can lead to inconsistencies where Pyrefly knows a function or class exist, but Pysa doesn't.

As a result, I believe Pyrefly should be responsible for generating the list of functions and classes, which is what the previous diffs do. However, we still need to generate unique fuly qualified name for those classes and functions, and this should still be Pysa's responsibility.

For that purpose, this diff introduces a "parent" attribute to functions and classes, which will be used to build fully qualified names on the Pysa side.

Reviewed By: tianhan0

Differential Revision: D79645266

fbshipit-source-id: d2c9a844f351fb4760bd45240bd56c4397414b03