Move `tuple_base` from `ClassMetadata` to `ClassBases`
Summary:
The `tuple_base` field in `ClassMetadata` requires knowing what kind of tuple a given base is, which in turn requires full typing info (including targs) on the base class type. We don't really want it to stay in `ClassMetadata` to avoid forming cycle with targs.

Let's move `tuple_base` to `ClassBases` instead so we don't have to worry about the cycle.

Reviewed By: yangdanny97, ndmitchell

Differential Revision: D79490220

fbshipit-source-id: 6ed9991e2f84a9f83d333df1412cff56decd71a0