Do not rely on targs outside of base class parsing within `class_metadata_of`
Summary:
With D79571092, we had made sure that any *downstream* of class metadata won't query class targs info. It still remains the case that the process of computing class metadata itself (namely, the `class_metadata_of()` method) still queries class targs info. The rest of the stack will be dedicated to eliminate that query as well.

`class_metadata_of()` is a big function. This diff makes it such that, instead of carrying around a list that contains `ClassType` (which contains targs info) around and only extract the `Class` object from it before returning, we make that list contain `Class` object directly instead, significantly reducing the scope in which targs info is available within the function. This prepares us for further refactors where we only need to worry about targs info in the very beginning of the `class_metadata_of()` function.

Reviewed By: ndmitchell

Differential Revision: D79594665

fbshipit-source-id: 2a17dbb1efb7948881252c1e6f5395f8a59ecba2