Inject typed dict fallback class to `ClassBases`
Summary: This is the final gap between the base classes computed by `ClassMetadata` and `ClassBases`. I was hoping that there's a way to only do it once but I wasn't able to figure out a good way of achieving it. So at the moment we'd have to duplicate the logic :(

Reviewed By: ndmitchell

Differential Revision: D79500458

fbshipit-source-id: 5d04b69c25acaec5f9d171be0045247be91de322