Simple code movement within `class_metadata_of`
Summary:
senator_we_shuffle

Pulling some blocks of code a bit up for better grouping. Technically speaking this is not 100% semantic-preserving for classes that are both dataclass and typeddict. But that seems to be an invalid and very edge case in general so it's not that important to 100% preserve our behavior there.

Reviewed By: ndmitchell

Differential Revision: D79594667

fbshipit-source-id: ffef85dd7b82f591928b56baeffb600c17cbffb6