Introduce new `ClassBaseType` key and binding
Summary:
Our goal here is to avoid class metadata computation to form a cycle with class tparams computation. But that's too hard to pull off today -- too many places in class metadata and its downstream depends on class tparams info.

What we can do instead is to break `ClassMetadata` down into 2 different bindings: one that depends on tparams info, and another one that doesn't. The general plan here is to turn `ClassMetadata` into the one that doesn't depend on tparams, and shovel all the checks and fields that do depend on tparams info into another binding.

This diff introduces that "another binding" which I named `ClassBases`, as the binding will hold the type info for all the base classes of a given class. Here I only did all the boilerplates without connecting them to the rest of the type checker. In subsequent diffs, we'll start to gradually move stuffs from `ClassMetadata` to `ClassBases` and chip away the current dependencies between metadata calc and tparams calc.

Reviewed By: ndmitchell

Differential Revision: D79490219

fbshipit-source-id: 07b590e0793bcded27b569516b628f324da572be