Instantiate generic functions in callable_infer, not as_call_target
Summary:
Before this diff, we created fresh vars and substituted them in when constructing a call target. This diff changes that, to preserve the tparams along with the function until we go to process the call with arguments.

This diff should preserve all existing behavior, but is arguably a bit simpler on its own. We don't have both `CallTarget` and `Target`, although we do gain a new `FunctionTarget`.

This change is prelude to contextual typing for calls to polymorphic functions based on expected return type. Ref https://github.com/facebook/pyrefly/issues/681

This change also makes it easier to fix https://github.com/facebook/pyrefly/issues/757

Reviewed By: grievejia

Differential Revision: D79598467

fbshipit-source-id: f58e7de6cc2b8bee581df73058c4f1ce50755823