Avoid querying synthesized field when synthesizing `__init__` for dataclass
Summary:
impacc_cycle

This diff gets rid of another trivial cycle we have: when synthesizing the `__init__()` method for dataclasses, we invoke `get_dataclass_member()` to grab all members in the dataclass (to use their type on the `__init__()` func signature). But under the hood `get_dataclass_member()` will query both synthesized and non-synthesized members in the dataclass, and when it queries the synthesized ones, `__init__()` would be part of it so we are going to hit the same binding again.

To break this cycle, let's make it so that it only queries the non-synthesized members, similar to what was done for namedtuples/typeddict in D78772415 / D78854638

Reviewed By: yangdanny97

Differential Revision: D79733913

fbshipit-source-id: a007eb7b60e8956782871ccdb828e95b3e55d752