Extend class metadata to include pydantic
Summary:
Context:

I've had a design discussion with grievejia and rchen152 based on the initial prototype (seen here D79305115), which had the following limitations:

- we do not use class index for pydantic binding key
- we have to check if binding exists
- also, I think another issue is that we will need to create many bindings based on future needs

After discussing,  we have converged on the following approach:

Instead of creating a separate pydantic binding and populating it directly at the AST level, modify the metadata for the class directly by extending the binding for the class.

I think this has advantages which are

- the issue that ndmitchell mentioned where we have to check if a binding exists will be addressed

- we can have a single binding and can access all the class information and populate it in that binding

The only risk we discussed is, what if we need some part of the AST that is not available on the class metadata binding level?

In that case, we agreed that we would extend the logic to propagate that data upwards.

Reviewed By: yangdanny97

Differential Revision: D79680955

fbshipit-source-id: 013a14898f732b6c3a3ca827df5771572620ea72