Add pydantic to special exports
Summary:
This diff is from our initial prototype.
We still need to add special exports here to determine if a ConfigDict is really a pydantic ConfigDict.

In this diff:

Know that we are importing from pydantic via special exports. Only then will we generate bindings for pydantic. We will do this next.

Reviewed By: ndmitchell

Differential Revision: D79305115

fbshipit-source-id: 9170816b7a3c8252f20cea22aeb515168e89a0c8