prioritize exit notifications
Summary:
D79177301 was bisected to lead to [this issue](https://github.com/facebook/pyrefly/issues/853) where the language server was failing to stop after being stopped.

## Background

A normal shutdown looks like this:

- shutdown request

[(no other messages are to be sent or it will error)](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#shutdown)

- exit message

We handle this [here](https://fburl.com/code/8hrgssfd) where we [take over the receiver until we have seen an exit](https://fburl.com/code/22bd0mcv).

## This change

This change prioritizes the exit notification in the lsp queue. This means `go-to definition, exit` is now handled in `exit, go-to definition` order.

this fixes the problem every time in local testing. there are still open questions, since as far as I can tell, the exit notification should not have been canceled before this diff

Why is this ok?

- after an exit is sent, if requests are still being processed, it seems safe to skip those requests
- shutdown requests (where shutdown must come before exit) are handled before this queue (so exits will not appear before a shutdown)

Another approach that seems to fix the issue is D79746701 (not sure why, open question). The fix in this diff seems better because it seems closer to the root of the problem.

### Open questions

- [stop()](https://fburl.com/6fkf2a61) (happens when we [end our language server](https://fburl.com/code/trs7eyk0)) calls [shutdown](https://github.com/microsoft/vscode-languageserver-node/blob/fc03016e0249448d120aa2fbf0f8ed7999ab5d50/client/src/common/client.ts#L1576C10-L1576C20), not only exit. Why doesn't our shutdown logic make us exit gracefully (this diff didn't change it)?
- why does awaiting exit fix this issue? D79746701
- why does this change fix the issue? prior to this diff, I don't think it was possible for exit notifications to be cancelled. could it be the case that vscode window reloads happened faster than we handled the exit if there were other mutations in between?

Reviewed By: SamChou19815

Differential Revision: D79759225

fbshipit-source-id: 149b7da1fc26f25e11bfc4c780e07ba3c2befcc6