Make find_one_part return remaining search roots
Summary:
# This diff...

Updates `find_one_part` to return the rest of the search roots, so that we can continue searching for an import if we want to attempt to find a better import alternative.

# This diff stack...

implements preference for `.pyi` files over `.py` over `.pyc` files in import finding, even if the import occurs later in the includes list.

Previous diffs set up for this by:
- creating `.py` and `.pyi` single file variants in `FindResult`
- removing `use_untyped_imports`

Future diffs will:
- perform a few more refactors to make the implementation easier (move `FindResult` conversion to `ModulePath` into a helper, create a comparator for `FindResult`, combine `find_module_in` approaches, since they don't differ by much anymore)
- implement the fallback finding behavior
- allow the LSP to specify if a `.py` or `.pyi` file is preferred

See bottom of diff stack for more details.

Reviewed By: grievejia

Differential Revision: D79490956

fbshipit-source-id: 482de40140f1c239180b28f26c0d14d29a369ba8