Search for better results when py or pyc is found in search_path
Summary:
# This diff...

creates a new `find_module_components` function, which is similar to how module finding currently works, but implements the fallback logic.

The basic idea is that: given a module or `module-stubs` first component, we attempt to find a match. If there's a match, but it's not the best option, we attempt to continue searching using the comparator from before. In upcoming diffs, we'll swap it out with the existing implementations in `find_in_search_path` and `find_in_site_package_path`.

# This diff stack...

implements preference for `.pyi` files over `.py` over `.pyc` files in import finding, even if the import occurs later in the includes list.

Previous diffs set up for this by:
- creating `.py` and `.pyi` single file variants in `FindResult`
- removing `use_untyped_imports`
- return remaining search roots from `find_one_part`
- create a `FindResult` comparator
- create method to convert `FindResult` to `ModulePath`

Future diffs will:
- implement the fallback finding behavior
- allow the LSP to specify if a `.py` or `.pyi` file is preferred

See bottom of diff stack for more details.

Reviewed By: grievejia

Differential Revision: D79560995

fbshipit-source-id: 36989167bbbb3d827802fc949e5697e998179115